Fix search: find text even if it is split across 2 pages
Note that I went for the simplest display of the page where the result was
found: if a hit is found spanning page 1 and 2, the display will say
`Found text at page 1'.  I hesitated implementing `Found text at page 1-2'
in this case, but was afraid of breaking someone's translations + it would
be more correct to have `Found text at pages 1-2' which would introduce a
new string to translate?