Add shortcut to manually add the current page as a navigation point
Sometimes while reading a document I need to scroll a few pages forward or backward (without clicking on any links or bookmarks) and quickly return to where I was. Now we can do this by pressing the shortcut ('y' as of this commit) before scrolling and then return by hitting backspace.