Remove BOM character from RTL lists
Unicode character 0xFEFF, although often presented along with the
"Arabic Presentation Forms-B" characters, is actually part of the
Specials characters, and represents the Byte Order Marker (BOM),
marking the [big]-endianness (when appearing at the beginning of a
file/stream) or a Zero Width No-Break Space (ZWNBSP) (when appearing
elsewhere).

Accordingly, it's presence in the Right-To-Left list of characters
would make texts seem more Right-to-Left oriented than they really are.

This fixes #1138, where in the table of contents of the submitted PDF
file there was a count of 57 RTL characters, all BOMs, against 4 LTR
characters, setting it's direction as Right-to-Left.