add option to exit when double cliking on the menu button
This commit add a boolean option MenuButtonDblClickExit.
When true, a double click on the menu button will cause SumatraPDF to
exit (the menu button is only shown when UseTabs is true).

This allows to close SumatraPDF like other Windows application using a
double click on left top icon of the title bar.
The double click duration is the one configured in Windows' settings.

Issue: #162