EnginePdf: fix memory leak when rendering is aborted
When the rendering (here in `fz_run_display_list`) is aborted, a memory
leak can occur as `fz_close_device` can throw in case of "items left on
stack in draw device".
This happens when zooming out and in.

When `fz_close_device` throws, `fz_drop_device` and `fz_drop_pixmap`
are not executed and so a memory leak occurs.

With this commit, if the rendering is aborted in a way that causes
`fz_close_device` to throw, `EnginePdf::RenderPage` function will
just return a null pointer as it would with any errors, without memory
leak.

Issue: #1393