Support dragging out a tab into a new window.
Ideally the separation of concerns should allow us to detach a WindowTab
from a MainWindow and to attach it to a different MainWindow, keeping
its DocController, engine, state of browsing, etc. intact. To help with
this, some light refactoring in this commit changes CreateNewTab() to
AddTabToWindow() and RemoveAndDeleteTab() to RemoveTab().

However, this doesn't work and crashes in a variety of places due to
duplicated data/concerns, most obviously win->ctrl vs tab->ctrl, but after
that's fixed, there's more around. To make a simple thing that'll work,
we ReloadDocument() in the newly created tab, thus losing page number,
browsing settings, etc., for now.