When dragged to a different window and released, add tab to that window
The change from ReloadDocument() to LoadDocument() is because the former
checks assumptions on win->ctrl and tab->ctrl that aren't true in our
novel use case. And anyway, LoadDocument with args->reuseTab is exactly
the right thing to do in our case; and is what ReloadDocument() would
have done if we hackily adjusted win->ctrl=nullptr for it.