Theme support: tabs, theme-switching key
The existing Theme structs have more than just colors for tabs
(pen width, whether or not to draw the X circle, etc.). Also the
stuff in AppColors.cpp doesn't do anything nontrivial with tab
colors (unlike main doc colors which several global settings can
influence). Taken together, it seems a good idea to let go the
AppColor intermediary for tab colors and get them from Theme directly.
This commit does that, plus add a Ctrl-Shift-D key to cycle the theme.
This is just for convenient testing, I don't know what the key should
be, and there probably should be a separate mechanism in the menu.

It is NOT a complete theming support, the main colors are trickier
and will come separately. I'm not yet sure whether to leave the AppColor
intermediary for them or incorporate the logic in Theme.cpp.

The stuff in Theme.cpp was a bit bitrot, but just a bit. Fixed a
compile error by changing char* name to const char*, and changed some
of the actual colors in the light theme to newer values in
AppColors.cpp.